Сделали AppError, вынесли коды ошибок

+++ src/services/task.service.js
@@ -1,67 +1,64 @@
 import db from '../models/index.js'
+import AppError from '../errors/AppError.js'
+import { ERROR_CODES } from '../constants/errorCodes.js'
+
 const { Task } = db
 
 class TaskService {
-	async getAllTasks() {
-		const tasks = await Task.findAll()
+	async getAllTasks(userId) {
+		const tasks = await Task.findAll({ where: { userId } })
 
-		if (!tasks) {
-			throw new Error('TASKS_NOT_FOUND')
+		if (!tasks.length) {
+			throw new AppError(ERROR_CODES.TASKS_NOT_FOUND, 404)
 		}
 
 		return tasks
 	}
 
 	async getTaskById(id) {
-		const task = await Task.findOne({
-			where: { id },
-		})
+		const task = await Task.findByPk(id)
 
 		if (!task) {
-			throw new Error('TASK_NOT_FOUND')
+			throw new AppError(ERROR_CODES.TASK_NOT_FOUND, 404)
 		}
 
 		return task
 	}
 
-	async createTask(title, description, userId) {
-		return await Task.create({ title, description, userId })
-	}
-
-	async updateTask(id, userId, body) {
-		const task = await Task.findOne({
-			where: { id, userId },
-		})
-
-		if (!task) {
-			throw new Error('TASK_NOT_FOUND')
+	async createTask({ title, description, userId }) {
+		if (!title) {
+			throw new AppError(ERROR_CODES.INCORRECT_DATA, 400)
 		}
 
-		await task.update(body)
-		return task
+		return Task.create({ title, description, userId })
+	}
+
+	async updateTask({ id, userId, body }) {
+		const task = await Task.findOne({ where: { id, userId } })
+
+		if (!task) {
+			throw new AppError(ERROR_CODES.TASK_NOT_FOUND, 404)
+		}
+
+		return task.update(body)
 	}
 
 	async deleteTask(id) {
 		const task = await Task.findByPk(id)
 
 		if (!task) {
-			throw new Error('TASK_NOT_FOUND')
+			throw new AppError(ERROR_CODES.TASK_NOT_FOUND, 404)
 		}
+
 		await task.destroy()
 	}
 
 	async deleteAllTasks(userId) {
-		const tasks = await Task.findAll({ where: userId })
+		const count = await Task.destroy({ where: { userId } })
 
-		if (!tasks) {
-			throw new Error('TASKS_NOT_FOUND')
+		if (!count) {
+			throw new AppError(ERROR_CODES.TASKS_NOT_FOUND, 404)
 		}
-
-		await Task.destroy({
-			where: {
-				userId: userId,
-			},
-		})
 	}
 }
